scripts/test-db/requirements.txt
psycopg2-binary
pandas
python-dotenv